Fix: Repair the consent issue when signup and add the specification of the social media when saving score
