+++ WebApp/Backend/models.py
@@ -25,6 +25,7 @@ class UsersScansResults(Base):
     src_posts_count = Column(Integer, nullable=False)
     src_highlights_count = Column(Integer, nullable=False)
     score = Column(Integer, nullable=False)
+    media = Column(String(255), nullable=True)
     created_at = Column(DateTime(timezone=True), server_default=func.now())
