+++ WebApp/Backend/routes/auth.py
@@ -1,0 +1,67 @@
+from fastapi import APIRouter, Depends, HTTPException
+from pydantic import BaseModel
+from sqlalchemy.orm import Session
+
+from db import get_db
+from models import User
+from utils import hash_password, verify_password
+from auth import create_access_token
+
+
+router = APIRouter(tags=["auth"])
+
+
+class UserCreate(BaseModel):
+    username: str
+    email: str
+    password: str
+    display_consent: bool
+    cgu: bool
+
+
+class UserLogin(BaseModel):
+    username: str
+    password: str
+
+
+@router.post("/register")
+def register(user: UserCreate, db: Session = Depends(get_db)):
+    db_user = db.query(User).filter(User.username == user.username).first()
+    if db_user:
+        raise HTTPException(status_code=400, detail="Username already registered")
+    if not user.cgu:
+        raise HTTPException(status_code=400, detail="Terms and conditions must be accepted")
+
+    hashed_pw = hash_password(user.password)
+    new_user = User(
+        username=user.username,
+        email=user.email,
+        password_hash=hashed_pw,
+        display_consent=user.display_consent,
+        cgu=user.cgu,
+    )
+    db.add(new_user)
+    db.commit()
+    db.refresh(new_user)
+
+    return {
+        "msg": "User created successfully",
+        "user_id": new_user.id,
+        "display_content": new_user.display_consent,
+    }
+
+
+@router.post("/login")
+def login(user: UserLogin, db: Session = Depends(get_db)):
+    db_user = db.query(User).filter(User.username == user.username).first()
+    if not db_user or not verify_password(user.password, db_user.password_hash):
+        raise HTTPException(status_code=401, detail="Invalid credentials")
+
+    token = create_access_token({"sub": db_user.username})
+
+    return {
+        "access_token": token,
+        "token_type": "bearer",
+        "user_id": db_user.id,
+        "display_content": db_user.display_consent,
+    }
